requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
beautifulsoup4==4.8.2
billiard==3.3.0.23
bs4==0.0.1
cached-property==1.5.1
cachetools==2.1.0
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cssselect==1.1.0
curlify==2.2.1
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.41.0
django-cors-headers==2.4.0
django-extensions==2.2.9
django-facebook==6.0.3
django-postgres-extensions==0.9.3
django-s3-sqlite==0.0.3
djangorestframework==3.11.0
facebook-business==5.0.4
facebook-py==0.2
facebook-scraper==0.1.8
facebookads==2.11.4
fake-useragent==0.1.11
google-api-python-client==1.7.4
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
googleads==14.1.0
gunicorn==20.0.4
httplib2==0.11.3
idna==2.8
importlib-metadata==1.6.0
isodate==0.6.0
isort==4.3.21
jsonschema==3.2.0
kombu==3.0.37
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
numpy==1.15.2
oauth2client==4.1.2
oauthlib==3.1.0
pandas==0.23.4
parse==1.15.0
Pillow==7.0.0
psycopg2-binary==2.8.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyee==7.0.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pymongo==3.10.1
pyppeteer==0.0.25
pyquery==1.4.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==3.13
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
rsa==3.4.2
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
suds-jurko==0.6
tqdm==4.43.0
typed-ast==1.4.1
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.8
w3lib==1.21.0
websockets==8.1
wrapt==1.11.2
xmltodict==0.11.0
zeep==3.1.0
zipp==3.1.0
